+++ src/adapters/gateways/user_storage_gateway.hpp
@@ -29,7 +29,7 @@ public:
 
 private slots:
     void proccessUserData(const QString& firstName, const QString& lastName,
-                          const QString& email, double usedBookStorage,
+                          const QString& email, long usedBookStorage,
                           const QJsonArray& tags);
 
     void reportFailureGettingUser();
