+++ src/adapters/interfaces/persistance/i_user_storage_access.hpp
@@ -33,7 +33,7 @@ public:
 
 signals:
     void userReady(const QString& firstName, const QString& lastName,
-                   const QString& email, double usedBookStorage,
+                   const QString& email, long usedBookStorage,
                    const QJsonArray& tags);
     void gettingUserFailed();
 };
