+++ src/application/interfaces/services/i_user_service.hpp
@@ -30,7 +30,7 @@ public:
     virtual QString getEmail() const = 0;
     virtual void setEmail(const QString& newEmail) = 0;
 
-    virtual double getUsedBookStorage() const = 0;
+    virtual long getUsedBookStorage() const = 0;
 
     virtual QImage getProfilePicture() const = 0;
     virtual void setProfilePicture(const QImage& image) = 0;
@@ -47,7 +47,7 @@ signals:
     void tagDeletionStarted(int index);
     void tagDeletionEnded();
     void tagsChanged(int index);
-    void usedBookStorageUpdated(double usedStorage);
+    void usedBookStorageUpdated(long usedStorage);
 
 public slots:
     virtual void setupUserData(const QString& token, const QString& email) = 0;
