+++ src/application/services/user_service.hpp
@@ -26,7 +26,7 @@ public:
     QString getEmail() const override;
     void setEmail(const QString& newEmail) override;
 
-    double getUsedBookStorage() const override;
+    long getUsedBookStorage() const override;
 
     QImage getProfilePicture() const override;
     void setProfilePicture(const QImage& image) override;
